fix/Minor fixes here and there (#37)
* Minor fix as per idex documentation

* fix typo in idex_active_order_tracker

* Comment out all legacy idex code as first step to delete it

* Prepare to delete legacy code
